(maint) Turns this into a proper component module
Instead of trying to manage the full stack with apache and all, this is
just a component module like it should be. Since we don't need the janky
multiple RVM environments and the weird symlinks anymore, there's no
need to force the end-to-end config. Instead, users should manage with a
profile.