(maint) add referer checking
This allows you to gist validated code. This adds an endpoint to
validate the referer and adds that link to the description.

Note: this doesn't actually do anything just yet. Browser security
policies don't send the referer header from HTTPS -> HTTP. But as soon
as we add HTTPS support to the validator, it will work.